Code significantly refactored
All the section-specific applications are now based off the core.scpt file, with the only change being a different parameter passed to the genPrompt() function ("news", "features", "arts", "sport").

In turn this function calls a generic page generator function, which eliminates a lot of repetition that was present in earlier versions.

This generation function lays the groundwork for a future custom edition generator.

The earlier versions are present in the "Legacy apps" folder if you want to see the evolution.