Fixed #5 Feb-March date bug, added Create generators.scpt
Like an idiot, I'd made it check if one string was greater than another.
theDate is now coerced to number and then checked against a number.

Create generators.scpt is essentially a makefile that automatically creates
run-only .app files for each desk, and a custom edition run-only script,
each with a custom icon. (I made it a while ago but have just added in
the icon automation and put it in here so it's under version control.)